linux-gen: socket_mmap: remove redundant header file
Remove odp_packet_socket.h as it is only used by
socket_mmap.c. Move relevant content to the c file.
