+++ platform/linux-generic/Makefile.am
@@ -111,7 +111,6 @@ noinst_HEADERS = \
 		  include/odp_packet_io_internal.h \
 		  include/odp_packet_io_ipc_internal.h \
 		  include/odp_packet_io_ring_internal.h \
-		  include/odp_packet_socket.h \
 		  include/odp_socket_common.h \
 		  include/odp_packet_io_stats_common.h \
 		  include/odp_packet_io_stats.h \
