--- platform/linux-generic/include/odp_packet_socket.h
@@ -1,62 +1,0 @@
-/* Copyright (c) 2013-2018, Linaro Limited
- * Copyright (c) 2013, Nokia Solutions and Networks
- * All rights reserved.
- *
- * SPDX-License-Identifier:     BSD-3-Clause
- */
-
-#ifndef ODP_PACKET_SOCKET_H
-#define ODP_PACKET_SOCKET_H
-
-#include <linux/if_packet.h>
-#include <linux/if_ether.h>
-#include <sys/socket.h>
-#include <string.h>
-#include <stddef.h>
-
-#include <odp/api/align.h>
-#include <odp/api/buffer.h>
-#include <odp/api/debug.h>
-#include <odp/api/pool.h>
-#include <odp/api/packet.h>
-#include <odp/api/packet_io.h>
-#include <odp/api/shared_memory.h>
-
-#include <linux/version.h>
-
-/*
- * Packet socket config:
- */
-
-/*
- * This makes sure that building for kernels older than 3.1 works
- * and a fanout requests fails (for invalid packet socket option)
- * in runtime if requested
- */
-#ifndef PACKET_FANOUT
-#define PACKET_FANOUT		18
-#define PACKET_FANOUT_HASH	0
-#endif /* PACKET_FANOUT */
-
-/** packet mmap ring */
-struct ring {
-	struct iovec *rd;
-	unsigned frame_num;
-	int rd_num;
-
-	odp_shm_t shm;
-	int sock;
-	int type;
-	int version;
-	uint8_t *mm_space;
-	size_t mm_len;
-	size_t rd_len;
-	int flen;
-
-	struct tpacket_req req;
-};
-
-ODP_STATIC_ASSERT(offsetof(struct ring, mm_space) <= ODP_CACHE_LINE_SIZE,
-		  "ERR_STRUCT_RING");
-
-#endif
